requirements.txt
# Core dependencies
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
networkx==3.1
matplotlib==3.7.2
seaborn==0.12.2

# Database
supabase==1.0.3
psycopg2-binary==2.9.7

# AI/ML dependencies for Phase 1
torch==2.0.1
torch-geometric==2.3.1
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0

# NLP dependencies
sentence-transformers==2.2.2
transformers==4.33.2
nltk==3.8.1

# Graph processing
igraph==0.10.6
python-igraph==0.10.6

# Data processing
openpyxl==3.1.2
xlrd==2.0.1
python-dateutil==2.8.2

# Security and authentication
bcrypt==4.0.1
PyJWT==2.8.0
cryptography==41.0.4

# Monitoring and logging
prometheus-client==0.17.1
structlog==23.1.0

# Async and concurrency
asyncio==3.4.3
aiohttp==3.8.5

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development tools
black==23.7.0
flake8==6.0.0
mypy==1.5.1

# Additional utilities
tqdm==4.65.0
click==8.1.7
rich==13.5.2

# Optional: Advanced features
optuna>=2.10.0  # Hyperparameter optimization
wandb>=0.12.0   # Experiment tracking 